requirements.txt
chex
dm_control
dm-haiku
jax  # See github for CUDA compatible installation.
jaxlib
optax
rltools[all] @ git+https://github.com/lkhromykh/rltools.git@dev
tensorflow_probability[jax]